refactor: rename service object
